Rename BookingCalendar to AdminDispatchCalendar and update usages
- Rename component to reflect dispatcher/admin ownership
- Update pages/calendar.vue import and tag
- Keep existing functionality; this is a naming + intent clarification only
